requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
httpx==0.27.2
celery==5.4.0
email-validator==2.2.0
bcrypt==4.0.1
pypdf==4.3.1
python-docx==1.1.2
httpx==0.27.2
python-multipart==0.0.9
